+++ mithril-client-cli/src/commands/cardano_block/certify.rs
@@ -41,8 +41,9 @@ impl CardanoBlocksCertifyCommand {
             ProgressOutputType::Tty
         };
         let progress_printer = ProgressPrinter::new(progress_output_type, 4);
+        let client_builder = context.setup_mithril_client_builder_with_fallback_genesis_key()?;
         let client = context
-            .setup_mithril_client_builder_with_fallback_genesis_key()?
+            .add_certificate_chain_cache(client_builder)?
             .add_feedback_receiver(Arc::new(IndicatifFeedbackReceiver::new(
                 progress_output_type,
                 logger.clone(),